+++ api-owner/src/main/java/com/wellmeet/reservation/dto/DayOfWeek.java
@@ -1,4 +1,4 @@
-package com.wellmeet.common;
+package com.wellmeet.reservation.dto;
 
 public enum DayOfWeek {
