+++ api-owner/src/main/java/com/wellmeet/restaurant/dto/UpdateOperatingHoursRequest.java
@@ -1,6 +1,6 @@
 package com.wellmeet.restaurant.dto;
 
-import com.wellmeet.common.DayOfWeek;
+import com.wellmeet.reservation.dto.DayOfWeek;
 import java.time.LocalTime;
 import java.util.List;
 import lombok.AllArgsConstructor;
